Add Event trait and EventId type
