+++ bee-common/bee-event-bus/Cargo.toml
@@ -3,7 +3,7 @@ name = "bee-event-bus"
 version = "0.1.0"
 authors = [ "IOTA Stiftung" ]
 edition = "2018"
-description = "A generic, type-safe and thread-safe event bus for arbitrary event types"
+description = "A generic, type-safe and thread-safe event bus for event types"
 readme = "README.md"
 repository = "https://github.com/iotaledger/bee"
 license = "Apache-2.0"
